76: Move the Slf4j Logger to the external package
