+++ app/src/test/groovy/gov/hhs/cdc/trustedintermediary/context/TestApplicationContext.groovy
@@ -1,7 +1,7 @@
 package gov.hhs.cdc.trustedintermediary.context
 
 import gov.hhs.cdc.trustedintermediary.wrappers.Logger
-import gov.hhs.cdc.trustedintermediary.wrappers.Slf4jLogger
+import gov.hhs.cdc.trustedintermediary.external.slf4j.Slf4jLogger
 
 /**
  * This test class resets the implementation registration in the ApplicationContext so different test cases can start on a clean slate.
