+++ app/src/test/groovy/gov/hhs/cdc/trustedintermediary/etor/order/OrderTest.groovy
@@ -6,7 +6,7 @@ import gov.hhs.cdc.trustedintermediary.context.TestApplicationContext
 import gov.hhs.cdc.trustedintermediary.wrappers.Formatter
 import gov.hhs.cdc.trustedintermediary.external.jackson.JacksonFormatter
 import gov.hhs.cdc.trustedintermediary.wrappers.Logger
-import gov.hhs.cdc.trustedintermediary.wrappers.Slf4jLogger
+import gov.hhs.cdc.trustedintermediary.external.slf4j.Slf4jLogger
 import spock.lang.Specification
 
 import java.time.LocalDateTime
